ARM: dts: msm: change CPR down threshold and quot adjustment for msm8926
The current VDD_APC CPR down thresdhold value of 1 makes the CPR controller
too sensitive to voltage ripple in the VDD_APC supply. This results in the
CPR UP and DOWN interrupts firing repeatedly in an alternating fashion.

Change the CPR down threshold to 2 and also change the CPR turbo and
nominal voltage corner target quotient adjustment values to 72 in order to
eliminate the undesirable CPR interrupt behavior.

CRs-fixed: 607472

Change-Id: I89b5c84c44c2c297f34023b177122441c65161a3
Signed-off-by: Ke Liu <keliu@codeaurora.org>
Signed-off-by: Manish Kumar <manishku@codeaurora.org>